Proportion of Reads and Replicates:

This method assigns taxa to the core if they account for some proportion of the total reads for the sequencing run and if they are present in at least x

In this example, a core taxa must account for 0.02 for the entire otu table and be present in at least 50

prop_reads_and_reps(
  otu_table,
  prop_reps = 0.5,
  prop_reads = 2e-04,
  taxa_as_rows = TRUE
)

Arguments

otu_table
a dataframe of OTUs where the first row is the OTU ID and column names refer to sites

prop_reps
percentage of replicates

prop_reads
percentage of reads

taxa_as_rows
logical data must be in a format where the taxa are rows and the sites are columns. The default value is TRUE, if FALSE data will be transposed for downstream analysis.

Value

the names of OTUs which meet the proportion of reads and replicate criteria

Examples

prop_reads_and_reps(arabidopsis)

#>   [1] "OTU_14834" "OTU_18567" "OTU_14402" "OTU_14822" "OTU_16757" "OTU_15154"
#>   [7] "OTU_7054"  "OTU_1196"  "OTU_18954" "OTU_14839" "OTU_13797" "OTU_11245"
#>  [13] "OTU_19706" "OTU_19511" "OTU_19108" "OTU_4555"  "OTU_19383" "OTU_18240"
#>  [19] "OTU_10942" "OTU_4035"  "OTU_7386"  "OTU_5475"  "OTU_15151" "OTU_18588"
#>  [25] "OTU_12650" "OTU_15877" "OTU_19673" "OTU_19937" "OTU_19482" "OTU_12773"
#>  [31] "OTU_3802"  "OTU_18407" "OTU_17901" "OTU_18945" "OTU_333"   "OTU_17465"
#>  [37] "OTU_19245" "OTU_11920" "OTU_14820" "OTU_2324"  "OTU_6893"  "OTU_14939"
#>  [43] "OTU_17455" "OTU_18209" "OTU_16101" "OTU_15808" "OTU_14817" "OTU_9086" 
#>  [49] "OTU_15529" "OTU_3497"  "OTU_4100"  "OTU_13953" "OTU_14928" "OTU_18468"
#>  [55] "OTU_14818" "OTU_17863" "OTU_14830" "OTU_14771" "OTU_14747" "OTU_352"  
#>  [61] "OTU_18761" "OTU_13788" "OTU_14293" "OTU_16786" "OTU_14813" "OTU_4362" 
#>  [67] "OTU_14450" "OTU_467"   "OTU_19189" "OTU_10317" "OTU_3327"  "OTU_14816"
#>  [73] "OTU_1602"  "OTU_3503"  "OTU_19740" "OTU_18891" "OTU_19859" "OTU_1460" 
#>  [79] "OTU_13230" "OTU_13580" "OTU_9309"  "OTU_3106"  "OTU_14431" "OTU_19297"
#>  [85] "OTU_13390" "OTU_6206"  "OTU_15169" "OTU_19047" "OTU_8809"  "OTU_14814"
#>  [91] "OTU_319"   "OTU_5635"  "OTU_15408" "OTU_10839" "OTU_1425"  "OTU_14819"
#>  [97] "OTU_14829" "OTU_1274"  "OTU_3149"  "OTU_2168"  "OTU_3613"  "OTU_13635"
#> [103] "OTU_14586" "OTU_19656" "OTU_16467" "OTU_20081" "OTU_9926"  "OTU_3821" 
#> [109] "OTU_14787" "OTU_8660"  "OTU_16338" "OTU_7961"  "OTU_7481"  "OTU_14836"
#> [115] "OTU_878"   "OTU_773"   "OTU_704"   "OTU_17537" "OTU_15348" "OTU_392"  
#> [121] "OTU_1089"  "OTU_18752" "OTU_17579" "OTU_18314" "OTU_19720" "OTU_6688" 
#> [127] "OTU_5826"  "OTU_11102" "OTU_18923" "OTU_11219" "OTU_14800" "OTU_8055" 
#> [133] "OTU_300"   "OTU_4897"  "OTU_13843" "OTU_14937" "OTU_487"   "OTU_16680"
#> [139] "OTU_19523" "OTU_5001"  "OTU_12924" "OTU_636"   "OTU_16273" "OTU_7554" 
#> [145] "OTU_11054" "OTU_12587" "OTU_14998" "OTU_1324"  "OTU_13065" "OTU_13822"
#> [151] "OTU_3256"  "OTU_14821" "OTU_5043"  "OTU_13052" "OTU_13524" "OTU_15516"
#> [157] "OTU_6144"  "OTU_14702" "OTU_19992" "OTU_15063" "OTU_13846" "OTU_7562" 
#> [163] "OTU_15527" "OTU_14857" "OTU_14790" "OTU_4507"  "OTU_2343"  "OTU_14493"
#> [169] "OTU_16330" "OTU_15255" "OTU_12383" "OTU_14803" "OTU_13874" "OTU_18190"
#> [175] "OTU_10284" "OTU_4678"  "OTU_18685" "OTU_8814"  "OTU_9818"  "OTU_14438"
#> [181] "OTU_15539" "OTU_14764" "OTU_4521"  "OTU_15592" "OTU_5098"  "OTU_18653"
#> [187] "OTU_15091" "OTU_15545" "OTU_12154" "OTU_12332" "OTU_15667" "OTU_19348"
#> [193] "OTU_11998" "OTU_16921" "OTU_14833" "OTU_8455"  "OTU_14841" "OTU_12192"
#> [199] "OTU_17113" "OTU_20120" "OTU_14845" "OTU_16572" "OTU_12042" "OTU_19221"
#> [205] "OTU_8518"  "OTU_18475" "OTU_130"   "OTU_17717" "OTU_14961" "OTU_3003" 
#> [211] "OTU_3891"  "OTU_18720" "OTU_1279"  "OTU_128"   "OTU_1902"  "OTU_14737"
#> [217] "OTU_14812" "OTU_7372"  "OTU_17861" "OTU_18269" "OTU_695"   "OTU_16355"
#> [223] "OTU_9648"  "OTU_15269" "OTU_18661" "OTU_14376" "OTU_12018" "OTU_56"   
#> [229] "OTU_16199" "OTU_13893" "OTU_19235" "OTU_14918" "OTU_19298" "OTU_6529" 
#> [235] "OTU_3915"  "OTU_15088" "OTU_1583"  "OTU_17800" "OTU_18044" "OTU_15242"
#> [241] "OTU_14909" "OTU_14907" "OTU_16745" "OTU_19484" "OTU_18470" "OTU_14915"
#> [247] "OTU_3448"  "OTU_12673" "OTU_2723"  "OTU_14881" "OTU_6011"  "OTU_13494"
#> [253] "OTU_19241" "OTU_9929"  "OTU_14952" "OTU_14870" "OTU_14777" "OTU_19335"
#> [259] "OTU_437"   "OTU_14835" "OTU_1966"  "OTU_16400" "OTU_16798" "OTU_15188"
#> [265] "OTU_3291"  "OTU_3286"  "OTU_10852" "OTU_14653" "OTU_14958" "OTU_14785"
#> [271] "OTU_631"   "OTU_2031"  "OTU_15463" "OTU_18756" "OTU_5138"  "OTU_14520"
#> [277] "OTU_14840" "OTU_9848"  "OTU_16369" "OTU_15126" "OTU_12463" "OTU_9793" 
#> [283] "OTU_11855" "OTU_3630"  "OTU_14956" "OTU_19405" "OTU_13290" "OTU_10594"
#> [289] "OTU_20135" "OTU_2108"  "OTU_4356"  "OTU_3152"  "OTU_3189"  "OTU_14158"
#> [295] "OTU_331"   "OTU_14865" "OTU_18594" "OTU_4382"  "OTU_10452" "OTU_12914"
#> [301] "OTU_15917" "OTU_1075"  "OTU_12578" "OTU_14837" "OTU_14342" "OTU_9590" 
#> [307] "OTU_377"   "OTU_4653"  "OTU_16845" "OTU_886"   "OTU_7730"  "OTU_19825"
#> [313] "OTU_19916" "OTU_229"   "OTU_14916" "OTU_14644" "OTU_19699" "OTU_15863"
#> [319] "OTU_12592" "OTU_2179"  "OTU_3093"  "OTU_14723" "OTU_17746" "OTU_14432"
#> [325] "OTU_15553" "OTU_9720"  "OTU_14613" "OTU_17301" "OTU_14745" "OTU_12282"
#> [331] "OTU_14755" "OTU_7820"  "OTU_10598" "OTU_1724"  "OTU_18404" "OTU_4225" 
#> [337] "OTU_14516" "OTU_15752" "OTU_10535" "OTU_1056"  "OTU_14556" "OTU_18542"
#> [343] "OTU_1538"  "OTU_7569"  "OTU_14060" "OTU_15597" "OTU_17931" "OTU_19214"
#> [349] "OTU_9906"  "OTU_18968" "OTU_19070" "OTU_14663" "OTU_14926" "OTU_16217"
#> [355] "OTU_4631"  "OTU_4142"  "OTU_13556" "OTU_14429" "OTU_7907"  "OTU_19728"
#> [361] "OTU_821"   "OTU_10219" "OTU_15214" "OTU_14784" "OTU_68"    "OTU_5670" 
#> [367] "OTU_3862"  "OTU_13428" "OTU_14924" "OTU_8839"  "OTU_3257"  "OTU_14945"
#> [373] "OTU_18495" "OTU_18662" "OTU_2173"  "OTU_1768"  "OTU_7536"  "OTU_20024"
#> [379] "OTU_13730" "OTU_14495" "OTU_14443" "OTU_13818" "OTU_18554" "OTU_14949"
#> [385] "OTU_14778" "OTU_6606"  "OTU_5365"  "OTU_14498" "OTU_1971"  "OTU_2898" 
#> [391] "OTU_14897" "OTU_14636" "OTU_13121" "OTU_9796"  "OTU_9506"  "OTU_14772"
#> [397] "OTU_14991" "OTU_19423" "OTU_8277"  "OTU_11173" "OTU_10733" "OTU_18631"
#> [403] "OTU_3011"  "OTU_4003"  "OTU_3015"  "OTU_2608"  "OTU_17549" "OTU_1831" 
#> [409] "OTU_12294" "OTU_18228" "OTU_255"   "OTU_7731"  "OTU_17480" "OTU_9425" 
#> [415] "OTU_14016" "OTU_13687" "OTU_14506" "OTU_15884" "OTU_16468" "OTU_13272"
#> [421] "OTU_14902" "OTU_2433"  "OTU_14073" "OTU_19681" "OTU_351"   "OTU_18580"
#> [427] "OTU_14399" "OTU_5543"  "OTU_14810" "OTU_7689"  "OTU_8307"  "OTU_5964" 
#> [433] "OTU_2544"  "OTU_4584"  "OTU_1250"  "OTU_18855" "OTU_15018" "OTU_3676" 
#> [439] "OTU_15746" "OTU_14592" "OTU_2820"  "OTU_3349"  "OTU_14034" "OTU_14807"
#> [445] "OTU_18814" "OTU_18886" "OTU_13997" "OTU_13067" "OTU_14368" "OTU_4991" 
#> [451] "OTU_14584" "OTU_16364" "OTU_5627"  "OTU_19960" "OTU_14802" "OTU_15743"
#> [457] "OTU_15582" "OTU_19075" "OTU_4516"  "OTU_15409" "OTU_14809" "OTU_19692"
#> [463] "OTU_14583" "OTU_16829" "OTU_18916" "OTU_8309"  "OTU_19222" "OTU_16195"
#> [469] "OTU_6925"  "OTU_14781" "OTU_5307"  "OTU_6361"  "OTU_18708" "OTU_12903"
#> [475] "OTU_10056" "OTU_2647"  "OTU_15513" "OTU_5532"  "OTU_14284" "OTU_1793" 
#> [481] "OTU_14541" "OTU_16316" "OTU_15741" "OTU_18910" "OTU_19330" "OTU_1278" 
#> [487] "OTU_11307" "OTU_19533" "OTU_5926"  "OTU_232"   "OTU_15026" "OTU_17741"
#> [493] "OTU_18237" "OTU_10819" "OTU_19396" "OTU_12857" "OTU_12023" "OTU_812"  
#> [499] "OTU_12022" "OTU_12955" "OTU_14067" "OTU_1064"  "OTU_3197"  "OTU_10487"
#> [505] "OTU_14875" "OTU_17930" "OTU_14713" "OTU_14979" "OTU_13614" "OTU_18556"
#> [511] "OTU_19919" "OTU_6064"  "OTU_9191"  "OTU_3498"  "OTU_19114" "OTU_11128"
#> [517] "OTU_580"   "OTU_14658" "OTU_1481"  "OTU_4286"  "OTU_5611"  "OTU_18422"
#> [523] "OTU_2901"  "OTU_14491" "OTU_1059"  "OTU_9892"  "OTU_15142" "OTU_13182"
#> [529] "OTU_321"   "OTU_485"   "OTU_11795" "OTU_3616"  "OTU_15632" "OTU_16865"
#> [535] "OTU_8316"  "OTU_6219"  "OTU_14925" "OTU_14774" "OTU_16212" "OTU_11616"
#> [541] "OTU_16055" "OTU_15334" "OTU_18292" "OTU_9984"  "OTU_700"   "OTU_16753"
#> [547] "OTU_12792" "OTU_12716" "OTU_19643" "OTU_15028" "OTU_14531" "OTU_15515"
#> [553] "OTU_14923" "OTU_2946"  "OTU_16428" "OTU_6608"  "OTU_20109" "OTU_474"  
#> [559] "OTU_7275"  "OTU_15710" "OTU_18137" "OTU_20073" "OTU_9150"  "OTU_18294"
#> [565] "OTU_14748" "OTU_4794"  "OTU_15359" "OTU_16440" "OTU_14887" "OTU_5471" 
#> [571] "OTU_6563"  "OTU_1651"  "OTU_14873" "OTU_11208" "OTU_247"   "OTU_12978"
#> [577] "OTU_14799" "OTU_11318" "OTU_15608" "OTU_14576" "OTU_14587" "OTU_14930"
#> [583] "OTU_9005"  "OTU_600"   "OTU_15783" "OTU_7647"  "OTU_19122" "OTU_2160" 
#> [589] "OTU_12175" "OTU_8404"  "OTU_4589"  "OTU_19049" "OTU_18249" "OTU_17571"
#> [595] "OTU_5345"  "OTU_8897"  "OTU_3728"  "OTU_3276"  "OTU_7800"  "OTU_7077" 
#> [601] "OTU_1899"  "OTU_15903" "OTU_14021" "OTU_7897"  "OTU_5226"  "OTU_15526"
#> [607] "OTU_5418"  "OTU_15062" "OTU_19576" "OTU_50"    "OTU_8041"  "OTU_12944"
#> [613] "OTU_12885" "OTU_14890" "OTU_14710" "OTU_18865" "OTU_612"   "OTU_10486"
#> [619] "OTU_3858"  "OTU_13848" "OTU_9312"  "OTU_1699"  "OTU_15680" "OTU_10684"
#> [625] "OTU_3316"  "OTU_12805" "OTU_2486"  "OTU_18655" "OTU_18259" "OTU_11763"
#> [631] "OTU_10061" "OTU_18015" "OTU_2379"  "OTU_2575"  "OTU_11346" "OTU_14526"
#> [637] "OTU_760"   "OTU_6802"  "OTU_2212"  "OTU_14904" "OTU_3864"  "OTU_1569" 
#> [643] "OTU_18672" "OTU_4221"  "OTU_17254" "OTU_12455" "OTU_14711" "OTU_3045" 
#> [649] "OTU_2996"  "OTU_15555" "OTU_17722" "OTU_14917" "OTU_14500" "OTU_5121" 
#> [655] "OTU_14570" "OTU_5031"  "OTU_14470" "OTU_17907" "OTU_18203" "OTU_8047" 
#> [661] "OTU_14149" "OTU_10116" "OTU_13106" "OTU_16159" "OTU_15440" "OTU_5091" 
#> [667] "OTU_14121" "OTU_9614"  "OTU_16164" "OTU_15333" "OTU_4676"  "OTU_12200"
#> [673] "OTU_13141" "OTU_12525" "OTU_551"   "OTU_1308"  "OTU_8855"  "OTU_14941"
#> [679] "OTU_16535" "OTU_12401" "OTU_12061" "OTU_18020" "OTU_14649" "OTU_3341" 
#> [685] "OTU_7232"  "OTU_10372" "OTU_6617"  "OTU_6454"  "OTU_1588"  "OTU_14173"
#> [691] "OTU_15770" "OTU_15266" "OTU_15717" "OTU_6510"  "OTU_5182"  "OTU_6869" 
#> [697] "OTU_13770" "OTU_4277"  "OTU_15379" "OTU_2527"  "OTU_14888" "OTU_12800"
#> [703] "OTU_15455" "OTU_8576"  "OTU_1138"  "OTU_5990"  "OTU_15252" "OTU_13406"
#> [709] "OTU_8988"  "OTU_16867" "OTU_11694" "OTU_12356" "OTU_3751"  "OTU_14878"
#> [715] "OTU_881"   "OTU_963"   "OTU_12637" "OTU_19362" "OTU_3779"  "OTU_11084"
#> [721] "OTU_16537" "OTU_19899" "OTU_18414" "OTU_12146" "OTU_12776" "OTU_4233" 
#> [727] "OTU_16248" "OTU_1291"  "OTU_16919" "OTU_13367" "OTU_16253" "OTU_18815"
#> [733] "OTU_14763" "OTU_427"   "OTU_618"   "OTU_10918" "OTU_14676" "OTU_3194" 
#> [739] "OTU_14536" "OTU_14524" "OTU_13074" "OTU_12241" "OTU_18952" "OTU_18309"
#> [745] "OTU_354"   "OTU_16145" "OTU_12654" "OTU_16734" "OTU_14577" "OTU_9112" 
#> [751] "OTU_10347" "OTU_3300"  "OTU_14514" "OTU_6875"  "OTU_4348"  "OTU_20067"
#> [757] "OTU_17553" "OTU_12088" "OTU_14667" "OTU_1183"  "OTU_14442" "OTU_19371"
#> [763] "OTU_16285" "OTU_18037" "OTU_14477" "OTU_4275"  "OTU_11673" "OTU_6055" 
#> [769] "OTU_19414" "OTU_12220" "OTU_15340" "OTU_12710" "OTU_10892" "OTU_13217"
#> [775] "OTU_12858" "OTU_15511" "OTU_12172" "OTU_19303" "OTU_15181" "OTU_2373" 
#> [781] "OTU_309"   "OTU_9838"  "OTU_10146" "OTU_11252" "OTU_16960" "OTU_12909"
#> [787] "OTU_15727" "OTU_3469"  "OTU_17839" "OTU_14899" "OTU_14561" "OTU_15229"
#> [793] "OTU_1370"  "OTU_9099"  "OTU_4378"  "OTU_10339" "OTU_6671"  "OTU_14645"
#> [799] "OTU_710"   "OTU_6101"  "OTU_5238"  "OTU_4384"  "OTU_3359"  "OTU_5240" 
#> [805] "OTU_15744" "OTU_14497" "OTU_10113" "OTU_11993" "OTU_14258" "OTU_2246" 
#> [811] "OTU_9460"  "OTU_12907" "OTU_17524" "OTU_9294"  "OTU_20052" "OTU_4104" 
#> [817] "OTU_3734"  "OTU_19889" "OTU_1132"  "OTU_4158"  "OTU_15521" "OTU_419"  
#> [823] "OTU_14501" "OTU_13075" "OTU_7471"  "OTU_18937" "OTU_4662"  "OTU_5131" 
#> [829] "OTU_2181"  "OTU_3242"  "OTU_14806" "OTU_5130"  "OTU_5467"  "OTU_9610" 
#> [835] "OTU_20026" "OTU_16816" "OTU_3183"  "OTU_6970"  "OTU_16060" "OTU_15082"
#> [841] "OTU_16679" "OTU_14440" "OTU_5432"  "OTU_6146"  "OTU_14301" "OTU_12751"
#> [847] "OTU_14503" "OTU_11985" "OTU_12913" "OTU_10574" "OTU_17366" "OTU_4986" 
#> [853] "OTU_264"   "OTU_9809"  "OTU_7315"  "OTU_14391" "OTU_20049" "OTU_9607" 
#> [859] "OTU_10890" "OTU_20025" "OTU_8850"  "OTU_14668" "OTU_17515" "OTU_13352"
#> [865] "OTU_14670" "OTU_20027" "OTU_778"   "OTU_19421" "OTU_11992" "OTU_14891"
#> [871] "OTU_14619" "OTU_2773"  "OTU_310"   "OTU_12846" "OTU_7287"  "OTU_11740"
#> [877] "OTU_11756" "OTU_6815"  "OTU_2239"  "OTU_5234"  "OTU_19046" "OTU_3261" 
#> [883] "OTU_1910"  "OTU_2896"  "OTU_18898" "OTU_1701"  "OTU_9805"  "OTU_16113"
#> [889] "OTU_15449" "OTU_268"   "OTU_18318" "OTU_6797"  "OTU_10621" "OTU_2708" 
#> [895] "OTU_7247"  "OTU_18548" "OTU_14426" "OTU_7702"  "OTU_7859"  "OTU_1798" 
#> [901] "OTU_3940"  "OTU_14455" "OTU_15620" "OTU_14449" "OTU_14741" "OTU_14716"
#> [907] "OTU_7527"