Proportion of Replicates

This method assigns taxa to the core if they account for some proportion of the total reads for the sequencing run and if they are present in at least x In this example, a core taxa must account for 0.01 for the entire otu table and be present in at least 50

prop_reps(otu_table)

Arguments

otu_table
a dataframe of OTUs where the first row is the OTU ID and column names refer to sites

Value

the names of OTUs which meet the proportion of replicate criteria

Examples

prop_reps(arabidopsis)

#>  [1] "OTU_10317" "OTU_10942" "OTU_11245" "OTU_11920" "OTU_1196"  "OTU_12192"
#>  [7] "OTU_1274"  "OTU_12773" "OTU_12924" "OTU_13052" "OTU_13390" "OTU_14293"
#> [13] "OTU_14431" "OTU_14586" "OTU_1460"  "OTU_14747" "OTU_14814" "OTU_14818"
#> [19] "OTU_14819" "OTU_14820" "OTU_14822" "OTU_14829" "OTU_14834" "OTU_14839"
#> [25] "OTU_14939" "OTU_15151" "OTU_15154" "OTU_15169" "OTU_15348" "OTU_15408"
#> [31] "OTU_15808" "OTU_15877" "OTU_16467" "OTU_16757" "OTU_16786" "OTU_17465"
#> [37] "OTU_18209" "OTU_18240" "OTU_18468" "OTU_18567" "OTU_18761" "OTU_18891"
#> [43] "OTU_18923" "OTU_18945" "OTU_18954" "OTU_19047" "OTU_19108" "OTU_19189"
#> [49] "OTU_19245" "OTU_19383" "OTU_19482" "OTU_19511" "OTU_19673" "OTU_19706"
#> [55] "OTU_19859" "OTU_3106"  "OTU_319"   "OTU_3327"  "OTU_3497"  "OTU_3503" 
#> [61] "OTU_352"   "OTU_3613"  "OTU_3802"  "OTU_3821"  "OTU_392"   "OTU_4035" 
#> [67] "OTU_4100"  "OTU_5475"  "OTU_5635"  "OTU_704"   "OTU_7054"  "OTU_7481" 
#> [73] "OTU_8660"  "OTU_8809"  "OTU_9086"  "OTU_9309"