Core Venn Diagram

Create a venn diagram using the output of core_methods to find which core taxa are included in which methods

core_venn(
  combined_taxa_data,
  category_names = c("Summation of \n Sequence Reads",
    "Proportion of \n Sequence Replicates",
    "Proportion of \n Sequence Reads and Replicates", "Hard Cut Off"),
  low = "white",
  high = "#7E191B"
)

# S3 method for default
core_venn(
  combined_taxa_data,
  category_names = c("Summation of \n Sequence Reads",
    "Proportion of \n Sequence Replicates",
    "Proportion of \n Sequence Reads and Replicates", "Hard Cut Off"),
  low = "white",
  high = "#7E191B"
)

# S3 method for core_methods
core_venn(
  combined_taxa_data,
  category_names = c("Summation of \n Sequence Reads",
    "Proportion of \n Sequence Replicates",
    "Proportion of \n Sequence Reads and Replicates", "Hard Cut Off"),
  low = "white",
  high = "#7E191B"
)

Arguments

combined_taxa_data
the taxa data of type core_methods

category_names
the names of the core inclusion methods.

low
the low hexedicmal color

high
the high hexidecimal color

Value

a venn diagram of the taxa included within each core method

Examples

if (FALSE) {
core_arabidopsis <- core_methods(arabidopsis)
core_venn(core_arabidopsis)
}