Hard Cut-Off

This method assigns taxa to the core if they are present in more than a pre-determined number of sites and have a total abidance greater than a pre-determined number of reads. In our example we set the minimum number of sites to 5 and the minimum number of reads to 25. Here we use the hard cut off described in Lundberg (2012), but realize this is any threshold.

hard_cutoff(otu_table, cutoff = 25, sites = 5, taxa_as_rows = TRUE)

Arguments

otu_table

a dataframe of OTUs where the first row is the OTU ID and column names refer to sites

cutoff

the threshhold minimum number of reads - defaults to 25

sites

the theshhold minimum number of sites - defaults to 5

taxa_as_rows

logical data must be in a format where the taxa are rows and the sites are columns. The default value is TRUE, if FALSE data will be transposed for downstream analysis.

Value

the names of OTUs which meet the hard cut off criteria

Examples

if (FALSE) {
hard_cutoff(arabidopsis)
}