Proportion of Replicates

This method assigns taxa to the core if they account for some proportion of the total reads for the sequencing run and if they are present in at least x In this example, a core taxa must account for 0.01 for the entire otu table and be present in at least 50

According to the mansucript it should be 0.5 x total reps. I hope we didnt do 10 x totalreps when calculating the number of taxa included etc.

prop_reps(otu_table, prop_reps = 0.5, taxa_as_rows = TRUE)

Arguments

otu_table
a dataframe of OTUs where the first row is the OTU ID and column names refer to sites

prop_reps
some definition

taxa_as_rows
logical data must be in a format where the taxa are rows and the sites are columns. The default value is TRUE, if FALSE data will be transposed for downstream analysis.

Value

the names of OTUs which meet the proportion of replicate criteria

Examples

prop_reps(arabidopsis)

#>    [1] "OTU_10000" "OTU_10011" "OTU_10030" "OTU_10042" "OTU_10049" "OTU_10056"
#>    [7] "OTU_10061" "OTU_1008"  "OTU_10097" "OTU_10099" "OTU_10108" "OTU_10110"
#>   [13] "OTU_10111" "OTU_10113" "OTU_10116" "OTU_10146" "OTU_10149" "OTU_10219"
#>   [19] "OTU_10226" "OTU_10241" "OTU_10284" "OTU_10285" "OTU_10314" "OTU_10317"
#>   [25] "OTU_1032"  "OTU_10328" "OTU_10329" "OTU_10335" "OTU_10339" "OTU_10343"
#>   [31] "OTU_10347" "OTU_10372" "OTU_10385" "OTU_10390" "OTU_10394" "OTU_10445"
#>   [37] "OTU_10452" "OTU_10453" "OTU_10486" "OTU_10487" "OTU_10504" "OTU_10535"
#>   [43] "OTU_10548" "OTU_1056"  "OTU_10574" "OTU_1059"  "OTU_10594" "OTU_10597"
#>   [49] "OTU_10598" "OTU_10621" "OTU_1064"  "OTU_10684" "OTU_1069"  "OTU_10694"
#>   [55] "OTU_10733" "OTU_10747" "OTU_1075"  "OTU_1076"  "OTU_10779" "OTU_10795"
#>   [61] "OTU_10796" "OTU_10819" "OTU_10839" "OTU_10841" "OTU_10852" "OTU_10874"
#>   [67] "OTU_10887" "OTU_1089"  "OTU_10890" "OTU_10892" "OTU_109"   "OTU_10901"
#>   [73] "OTU_10908" "OTU_10918" "OTU_10924" "OTU_10930" "OTU_10942" "OTU_10957"
#>   [79] "OTU_1097"  "OTU_10970" "OTU_11051" "OTU_11054" "OTU_11057" "OTU_11059"
#>   [85] "OTU_11084" "OTU_11097" "OTU_111"   "OTU_11101" "OTU_11102" "OTU_11128"
#>   [91] "OTU_11145" "OTU_11157" "OTU_11173" "OTU_112"   "OTU_11208" "OTU_11218"
#>   [97] "OTU_11219" "OTU_1122"  "OTU_11223" "OTU_11237" "OTU_11245" "OTU_11249"
#>  [103] "OTU_11252" "OTU_113"   "OTU_11307" "OTU_11318" "OTU_1132"  "OTU_11346"
#>  [109] "OTU_11370" "OTU_11375" "OTU_1138"  "OTU_11395" "OTU_11406" "OTU_11443"
#>  [115] "OTU_11492" "OTU_11510" "OTU_11513" "OTU_11537" "OTU_11614" "OTU_11616"
#>  [121] "OTU_11650" "OTU_11666" "OTU_11673" "OTU_11694" "OTU_117"   "OTU_11731"
#>  [127] "OTU_11740" "OTU_11749" "OTU_11751" "OTU_11756" "OTU_11763" "OTU_11766"
#>  [133] "OTU_11767" "OTU_11772" "OTU_11795" "OTU_11806" "OTU_1183"  "OTU_11855"
#>  [139] "OTU_11875" "OTU_11901" "OTU_11920" "OTU_11941" "OTU_1196"  "OTU_11974"
#>  [145] "OTU_11979" "OTU_11985" "OTU_11992" "OTU_11993" "OTU_11994" "OTU_11998"
#>  [151] "OTU_12000" "OTU_1201"  "OTU_12011" "OTU_12017" "OTU_12018" "OTU_12022"
#>  [157] "OTU_12023" "OTU_12040" "OTU_12042" "OTU_12056" "OTU_12061" "OTU_12078"
#>  [163] "OTU_12087" "OTU_12088" "OTU_12107" "OTU_12108" "OTU_12112" "OTU_12138"
#>  [169] "OTU_12143" "OTU_12144" "OTU_12146" "OTU_12154" "OTU_12172" "OTU_12173"
#>  [175] "OTU_12175" "OTU_12181" "OTU_12186" "OTU_12188" "OTU_12192" "OTU_12198"
#>  [181] "OTU_12200" "OTU_12220" "OTU_12223" "OTU_12225" "OTU_12241" "OTU_12280"
#>  [187] "OTU_12282" "OTU_12294" "OTU_12295" "OTU_12308" "OTU_1231"  "OTU_12315"
#>  [193] "OTU_12332" "OTU_12354" "OTU_12356" "OTU_12383" "OTU_12401" "OTU_12405"
#>  [199] "OTU_12453" "OTU_12455" "OTU_12462" "OTU_12463" "OTU_12465" "OTU_12471"
#>  [205] "OTU_12497" "OTU_1250"  "OTU_12502" "OTU_12524" "OTU_12525" "OTU_12528"
#>  [211] "OTU_12529" "OTU_12534" "OTU_12537" "OTU_12541" "OTU_12544" "OTU_12551"
#>  [217] "OTU_12576" "OTU_12578" "OTU_12587" "OTU_12590" "OTU_12592" "OTU_12617"
#>  [223] "OTU_12626" "OTU_12637" "OTU_12638" "OTU_12650" "OTU_12653" "OTU_12654"
#>  [229] "OTU_12665" "OTU_12669" "OTU_12673" "OTU_12681" "OTU_12685" "OTU_12687"
#>  [235] "OTU_12693" "OTU_12695" "OTU_12708" "OTU_12709" "OTU_12710" "OTU_12716"
#>  [241] "OTU_1274"  "OTU_12740" "OTU_12751" "OTU_12759" "OTU_12760" "OTU_12773"
#>  [247] "OTU_12774" "OTU_12776" "OTU_1278"  "OTU_1279"  "OTU_12792" "OTU_128"  
#>  [253] "OTU_1280"  "OTU_12800" "OTU_12805" "OTU_12830" "OTU_12832" "OTU_12837"
#>  [259] "OTU_12846" "OTU_1285"  "OTU_12852" "OTU_12857" "OTU_12858" "OTU_1286" 
#>  [265] "OTU_12873" "OTU_12877" "OTU_12885" "OTU_12897" "OTU_12903" "OTU_12907"
#>  [271] "OTU_12909" "OTU_1291"  "OTU_12913" "OTU_12914" "OTU_12924" "OTU_12940"
#>  [277] "OTU_12944" "OTU_12949" "OTU_12955" "OTU_12976" "OTU_12978" "OTU_12980"
#>  [283] "OTU_12984" "OTU_12994" "OTU_12995" "OTU_130"   "OTU_13008" "OTU_13045"
#>  [289] "OTU_13050" "OTU_13052" "OTU_13065" "OTU_13067" "OTU_13071" "OTU_13074"
#>  [295] "OTU_13075" "OTU_1308"  "OTU_13106" "OTU_13121" "OTU_13133" "OTU_13141"
#>  [301] "OTU_13172" "OTU_13182" "OTU_13184" "OTU_13189" "OTU_1319"  "OTU_13213"
#>  [307] "OTU_13217" "OTU_13230" "OTU_1324"  "OTU_13241" "OTU_13243" "OTU_13244"
#>  [313] "OTU_13262" "OTU_13271" "OTU_13272" "OTU_13278" "OTU_1328"  "OTU_13290"
#>  [319] "OTU_13293" "OTU_13301" "OTU_13306" "OTU_13327" "OTU_13339" "OTU_13345"
#>  [325] "OTU_13346" "OTU_13348" "OTU_13352" "OTU_13360" "OTU_13367" "OTU_13379"
#>  [331] "OTU_13388" "OTU_1339"  "OTU_13390" "OTU_13401" "OTU_13406" "OTU_13410"
#>  [337] "OTU_13418" "OTU_13428" "OTU_13463" "OTU_1348"  "OTU_13493" "OTU_13494"
#>  [343] "OTU_13515" "OTU_13519" "OTU_13524" "OTU_13528" "OTU_13529" "OTU_1355" 
#>  [349] "OTU_13556" "OTU_13580" "OTU_136"   "OTU_13601" "OTU_13605" "OTU_1361" 
#>  [355] "OTU_13613" "OTU_13614" "OTU_1362"  "OTU_13635" "OTU_13655" "OTU_13666"
#>  [361] "OTU_13679" "OTU_13683" "OTU_13687" "OTU_13697" "OTU_1370"  "OTU_13712"
#>  [367] "OTU_1372"  "OTU_13723" "OTU_13730" "OTU_13734" "OTU_13770" "OTU_13779"
#>  [373] "OTU_13788" "OTU_13797" "OTU_13807" "OTU_13811" "OTU_13818" "OTU_13822"
#>  [379] "OTU_13838" "OTU_13839" "OTU_13843" "OTU_13846" "OTU_13848" "OTU_13874"
#>  [385] "OTU_13893" "OTU_13912" "OTU_1392"  "OTU_13947" "OTU_13949" "OTU_13953"
#>  [391] "OTU_13985" "OTU_13997" "OTU_1400"  "OTU_14016" "OTU_14021" "OTU_14034"
#>  [397] "OTU_14036" "OTU_14044" "OTU_14047" "OTU_14057" "OTU_14060" "OTU_14063"
#>  [403] "OTU_14067" "OTU_14070" "OTU_14073" "OTU_14092" "OTU_14093" "OTU_14101"
#>  [409] "OTU_14115" "OTU_14121" "OTU_14127" "OTU_14135" "OTU_14139" "OTU_1414" 
#>  [415] "OTU_14149" "OTU_14158" "OTU_14170" "OTU_14172" "OTU_14173" "OTU_1418" 
#>  [421] "OTU_14181" "OTU_14185" "OTU_14203" "OTU_14205" "OTU_14227" "OTU_14230"
#>  [427] "OTU_1425"  "OTU_14258" "OTU_14268" "OTU_14279" "OTU_14282" "OTU_14284"
#>  [433] "OTU_14286" "OTU_14293" "OTU_14294" "OTU_14301" "OTU_14321" "OTU_14337"
#>  [439] "OTU_14339" "OTU_14342" "OTU_14363" "OTU_14368" "OTU_14376" "OTU_14384"
#>  [445] "OTU_14386" "OTU_14391" "OTU_14399" "OTU_14402" "OTU_14418" "OTU_14424"
#>  [451] "OTU_14426" "OTU_14428" "OTU_14429" "OTU_14431" "OTU_14432" "OTU_14437"
#>  [457] "OTU_14438" "OTU_14440" "OTU_14442" "OTU_14443" "OTU_14444" "OTU_14446"
#>  [463] "OTU_14447" "OTU_14449" "OTU_14450" "OTU_14455" "OTU_14461" "OTU_14463"
#>  [469] "OTU_14464" "OTU_14469" "OTU_14470" "OTU_14473" "OTU_14475" "OTU_14477"
#>  [475] "OTU_14480" "OTU_14484" "OTU_14489" "OTU_14490" "OTU_14491" "OTU_14493"
#>  [481] "OTU_14495" "OTU_14497" "OTU_14498" "OTU_14499" "OTU_145"   "OTU_1450" 
#>  [487] "OTU_14500" "OTU_14501" "OTU_14503" "OTU_14506" "OTU_14509" "OTU_14514"
#>  [493] "OTU_14516" "OTU_14520" "OTU_14524" "OTU_14525" "OTU_14526" "OTU_14531"
#>  [499] "OTU_14536" "OTU_14538" "OTU_1454"  "OTU_14541" "OTU_14545" "OTU_14548"
#>  [505] "OTU_14556" "OTU_14560" "OTU_14561" "OTU_14563" "OTU_1457"  "OTU_14570"
#>  [511] "OTU_14576" "OTU_14577" "OTU_14579" "OTU_14582" "OTU_14583" "OTU_14584"
#>  [517] "OTU_14586" "OTU_14587" "OTU_14592" "OTU_14593" "OTU_1460"  "OTU_14603"
#>  [523] "OTU_14604" "OTU_14606" "OTU_14612" "OTU_14613" "OTU_14616" "OTU_14618"
#>  [529] "OTU_14619" "OTU_14620" "OTU_14621" "OTU_14626" "OTU_14628" "OTU_14632"
#>  [535] "OTU_14635" "OTU_14636" "OTU_14637" "OTU_14641" "OTU_14643" "OTU_14644"
#>  [541] "OTU_14645" "OTU_14649" "OTU_14650" "OTU_14653" "OTU_14654" "OTU_14658"
#>  [547] "OTU_14659" "OTU_14660" "OTU_14663" "OTU_14666" "OTU_14667" "OTU_14668"
#>  [553] "OTU_14670" "OTU_14672" "OTU_14676" "OTU_14679" "OTU_14686" "OTU_14687"
#>  [559] "OTU_14692" "OTU_14694" "OTU_14702" "OTU_14710" "OTU_14711" "OTU_14713"
#>  [565] "OTU_14714" "OTU_14715" "OTU_14716" "OTU_14719" "OTU_14720" "OTU_14721"
#>  [571] "OTU_14722" "OTU_14723" "OTU_14725" "OTU_14729" "OTU_14731" "OTU_14733"
#>  [577] "OTU_14737" "OTU_14739" "OTU_1474"  "OTU_14741" "OTU_14743" "OTU_14744"
#>  [583] "OTU_14745" "OTU_14747" "OTU_14748" "OTU_14750" "OTU_14755" "OTU_14762"
#>  [589] "OTU_14763" "OTU_14764" "OTU_14766" "OTU_14770" "OTU_14771" "OTU_14772"
#>  [595] "OTU_14774" "OTU_14777" "OTU_14778" "OTU_14781" "OTU_14784" "OTU_14785"
#>  [601] "OTU_14786" "OTU_14787" "OTU_14790" "OTU_14791" "OTU_14799" "OTU_14800"
#>  [607] "OTU_14802" "OTU_14803" "OTU_14804" "OTU_14805" "OTU_14806" "OTU_14807"
#>  [613] "OTU_14808" "OTU_14809" "OTU_1481"  "OTU_14810" "OTU_14812" "OTU_14813"
#>  [619] "OTU_14814" "OTU_14815" "OTU_14816" "OTU_14817" "OTU_14818" "OTU_14819"
#>  [625] "OTU_14820" "OTU_14821" "OTU_14822" "OTU_14824" "OTU_14826" "OTU_14829"
#>  [631] "OTU_14830" "OTU_14832" "OTU_14833" "OTU_14834" "OTU_14835" "OTU_14836"
#>  [637] "OTU_14837" "OTU_14839" "OTU_14840" "OTU_14841" "OTU_14842" "OTU_14844"
#>  [643] "OTU_14845" "OTU_14848" "OTU_14850" "OTU_14854" "OTU_14857" "OTU_14863"
#>  [649] "OTU_14864" "OTU_14865" "OTU_14868" "OTU_14870" "OTU_14871" "OTU_14872"
#>  [655] "OTU_14873" "OTU_14875" "OTU_14878" "OTU_14881" "OTU_14882" "OTU_14883"
#>  [661] "OTU_14887" "OTU_14888" "OTU_14889" "OTU_14890" "OTU_14891" "OTU_14895"
#>  [667] "OTU_14897" "OTU_14898" "OTU_14899" "OTU_14901" "OTU_14902" "OTU_14904"
#>  [673] "OTU_14906" "OTU_14907" "OTU_14909" "OTU_1491"  "OTU_14910" "OTU_14911"
#>  [679] "OTU_14915" "OTU_14916" "OTU_14917" "OTU_14918" "OTU_14922" "OTU_14923"
#>  [685] "OTU_14924" "OTU_14925" "OTU_14926" "OTU_14928" "OTU_14930" "OTU_14933"
#>  [691] "OTU_14936" "OTU_14937" "OTU_14939" "OTU_14941" "OTU_14945" "OTU_14947"
#>  [697] "OTU_14948" "OTU_14949" "OTU_14952" "OTU_14956" "OTU_14958" "OTU_14961"
#>  [703] "OTU_14966" "OTU_14975" "OTU_14979" "OTU_14986" "OTU_14991" "OTU_14993"
#>  [709] "OTU_14998" "OTU_14999" "OTU_15005" "OTU_15007" "OTU_15008" "OTU_15010"
#>  [715] "OTU_15017" "OTU_15018" "OTU_15021" "OTU_15026" "OTU_15028" "OTU_1503" 
#>  [721] "OTU_15041" "OTU_15049" "OTU_15050" "OTU_15052" "OTU_15054" "OTU_15058"
#>  [727] "OTU_15059" "OTU_15062" "OTU_15063" "OTU_15068" "OTU_15082" "OTU_15083"
#>  [733] "OTU_15088" "OTU_15090" "OTU_15091" "OTU_15105" "OTU_15106" "OTU_15107"
#>  [739] "OTU_15121" "OTU_15126" "OTU_15134" "OTU_15142" "OTU_15144" "OTU_15150"
#>  [745] "OTU_15151" "OTU_15154" "OTU_15168" "OTU_15169" "OTU_15178" "OTU_15181"
#>  [751] "OTU_15182" "OTU_15188" "OTU_15201" "OTU_15202" "OTU_15210" "OTU_15214"
#>  [757] "OTU_15229" "OTU_15230" "OTU_15231" "OTU_15235" "OTU_15240" "OTU_15242"
#>  [763] "OTU_15245" "OTU_15248" "OTU_15251" "OTU_15252" "OTU_15255" "OTU_15257"
#>  [769] "OTU_15266" "OTU_15267" "OTU_15269" "OTU_15283" "OTU_15289" "OTU_15291"
#>  [775] "OTU_15296" "OTU_1532"  "OTU_15325" "OTU_15333" "OTU_15334" "OTU_15337"
#>  [781] "OTU_15340" "OTU_15343" "OTU_15348" "OTU_15350" "OTU_15351" "OTU_15359"
#>  [787] "OTU_15362" "OTU_15369" "OTU_15377" "OTU_15379" "OTU_1538"  "OTU_15382"
#>  [793] "OTU_15385" "OTU_15387" "OTU_15388" "OTU_15404" "OTU_15408" "OTU_15409"
#>  [799] "OTU_15423" "OTU_15425" "OTU_15430" "OTU_15435" "OTU_15439" "OTU_15440"
#>  [805] "OTU_15442" "OTU_15444" "OTU_15445" "OTU_15446" "OTU_15449" "OTU_15450"
#>  [811] "OTU_15452" "OTU_15455" "OTU_15458" "OTU_15460" "OTU_15463" "OTU_15465"
#>  [817] "OTU_15480" "OTU_15487" "OTU_15497" "OTU_15503" "OTU_15504" "OTU_15511"
#>  [823] "OTU_15513" "OTU_15515" "OTU_15516" "OTU_15518" "OTU_15521" "OTU_15524"
#>  [829] "OTU_15526" "OTU_15527" "OTU_15529" "OTU_15533" "OTU_15539" "OTU_15540"
#>  [835] "OTU_15542" "OTU_15543" "OTU_15545" "OTU_15553" "OTU_15555" "OTU_15559"
#>  [841] "OTU_15560" "OTU_15566" "OTU_15567" "OTU_15569" "OTU_15571" "OTU_15572"
#>  [847] "OTU_15573" "OTU_15582" "OTU_15592" "OTU_15597" "OTU_1560"  "OTU_15608"
#>  [853] "OTU_15610" "OTU_15613" "OTU_15615" "OTU_15616" "OTU_15617" "OTU_15620"
#>  [859] "OTU_15621" "OTU_15622" "OTU_15629" "OTU_15632" "OTU_15638" "OTU_15639"
#>  [865] "OTU_15645" "OTU_1565"  "OTU_15652" "OTU_15653" "OTU_15656" "OTU_15658"
#>  [871] "OTU_15661" "OTU_15667" "OTU_15671" "OTU_15673" "OTU_15678" "OTU_15679"
#>  [877] "OTU_15680" "OTU_15682" "OTU_15683" "OTU_1569"  "OTU_15690" "OTU_157"  
#>  [883] "OTU_15704" "OTU_15710" "OTU_15717" "OTU_15719" "OTU_15727" "OTU_15732"
#>  [889] "OTU_15737" "OTU_1574"  "OTU_15741" "OTU_15742" "OTU_15743" "OTU_15744"
#>  [895] "OTU_15745" "OTU_15746" "OTU_15748" "OTU_1575"  "OTU_15751" "OTU_15752"
#>  [901] "OTU_15755" "OTU_15756" "OTU_15759" "OTU_15761" "OTU_15762" "OTU_15767"
#>  [907] "OTU_15770" "OTU_15771" "OTU_15772" "OTU_15774" "OTU_15777" "OTU_15783"
#>  [913] "OTU_15792" "OTU_15796" "OTU_15801" "OTU_15802" "OTU_15808" "OTU_15809"
#>  [919] "OTU_15827" "OTU_1583"  "OTU_15844" "OTU_15857" "OTU_15863" "OTU_15877"
#>  [925] "OTU_1588"  "OTU_15884" "OTU_15891" "OTU_15903" "OTU_15913" "OTU_15914"
#>  [931] "OTU_15917" "OTU_15924" "OTU_15957" "OTU_15962" "OTU_15965" "OTU_15994"
#>  [937] "OTU_16007" "OTU_16018" "OTU_1602"  "OTU_16032" "OTU_16055" "OTU_16059"
#>  [943] "OTU_16060" "OTU_16076" "OTU_16083" "OTU_16098" "OTU_16101" "OTU_1611" 
#>  [949] "OTU_16113" "OTU_16145" "OTU_16149" "OTU_16159" "OTU_16164" "OTU_16167"
#>  [955] "OTU_16174" "OTU_16179" "OTU_16195" "OTU_16199" "OTU_16207" "OTU_16212"
#>  [961] "OTU_16214" "OTU_16216" "OTU_16217" "OTU_16242" "OTU_16248" "OTU_16253"
#>  [967] "OTU_16259" "OTU_16264" "OTU_16273" "OTU_16280" "OTU_16285" "OTU_163"  
#>  [973] "OTU_16316" "OTU_16330" "OTU_16338" "OTU_16348" "OTU_16349" "OTU_16355"
#>  [979] "OTU_16364" "OTU_16369" "OTU_16370" "OTU_16400" "OTU_16406" "OTU_16428"
#>  [985] "OTU_16432" "OTU_16440" "OTU_1645"  "OTU_16454" "OTU_16467" "OTU_16468"
#>  [991] "OTU_16492" "OTU_1651"  "OTU_16513" "OTU_16535" "OTU_16537" "OTU_16564"
#>  [997] "OTU_16572" "OTU_16580" "OTU_16618" "OTU_16657" "OTU_16665" "OTU_16676"
#> [1003] "OTU_16679" "OTU_16680" "OTU_16686" "OTU_1669"  "OTU_16699" "OTU_16702"
#> [1009] "OTU_16734" "OTU_16743" "OTU_16745" "OTU_16748" "OTU_16753" "OTU_16757"
#> [1015] "OTU_16786" "OTU_16798" "OTU_16816" "OTU_16827" "OTU_16829" "OTU_16845"
#> [1021] "OTU_16850" "OTU_16851" "OTU_16857" "OTU_16865" "OTU_16867" "OTU_16869"
#> [1027] "OTU_16873" "OTU_1688"  "OTU_16895" "OTU_169"   "OTU_16912" "OTU_16919"
#> [1033] "OTU_16921" "OTU_1694"  "OTU_16942" "OTU_16946" "OTU_16960" "OTU_16961"
#> [1039] "OTU_16974" "OTU_1699"  "OTU_16994" "OTU_17000" "OTU_17002" "OTU_17009"
#> [1045] "OTU_1701"  "OTU_17026" "OTU_17030" "OTU_17041" "OTU_1711"  "OTU_17113"
#> [1051] "OTU_17175" "OTU_17214" "OTU_17219" "OTU_17220" "OTU_1724"  "OTU_17254"
#> [1057] "OTU_17275" "OTU_17285" "OTU_17292" "OTU_1730"  "OTU_17301" "OTU_17311"
#> [1063] "OTU_17336" "OTU_17362" "OTU_17366" "OTU_17382" "OTU_17407" "OTU_1745" 
#> [1069] "OTU_17455" "OTU_17465" "OTU_17474" "OTU_17480" "OTU_17515" "OTU_17524"
#> [1075] "OTU_17537" "OTU_17542" "OTU_17545" "OTU_17549" "OTU_17553" "OTU_17558"
#> [1081] "OTU_17571" "OTU_17572" "OTU_17579" "OTU_17598" "OTU_17602" "OTU_17607"
#> [1087] "OTU_1761"  "OTU_1762"  "OTU_17645" "OTU_17658" "OTU_1768"  "OTU_17684"
#> [1093] "OTU_17687" "OTU_17717" "OTU_17722" "OTU_17741" "OTU_17746" "OTU_17749"
#> [1099] "OTU_1777"  "OTU_17776" "OTU_17777" "OTU_17800" "OTU_17825" "OTU_17834"
#> [1105] "OTU_17839" "OTU_1784"  "OTU_17843" "OTU_17846" "OTU_17861" "OTU_17863"
#> [1111] "OTU_17870" "OTU_17877" "OTU_17881" "OTU_1789"  "OTU_17897" "OTU_17901"
#> [1117] "OTU_17907" "OTU_1793"  "OTU_17930" "OTU_17931" "OTU_17968" "OTU_1798" 
#> [1123] "OTU_17983" "OTU_18001" "OTU_18008" "OTU_18015" "OTU_18017" "OTU_18020"
#> [1129] "OTU_18037" "OTU_18044" "OTU_18045" "OTU_18056" "OTU_18086" "OTU_18088"
#> [1135] "OTU_18097" "OTU_18104" "OTU_18105" "OTU_18123" "OTU_18131" "OTU_18137"
#> [1141] "OTU_18139" "OTU_18151" "OTU_18162" "OTU_18190" "OTU_18203" "OTU_18205"
#> [1147] "OTU_18209" "OTU_18228" "OTU_18237" "OTU_18240" "OTU_18249" "OTU_18259"
#> [1153] "OTU_18269" "OTU_18275" "OTU_18277" "OTU_18283" "OTU_18292" "OTU_18294"
#> [1159] "OTU_18295" "OTU_18297" "OTU_18301" "OTU_18309" "OTU_1831"  "OTU_18313"
#> [1165] "OTU_18314" "OTU_18318" "OTU_18343" "OTU_18363" "OTU_18393" "OTU_18404"
#> [1171] "OTU_18407" "OTU_1841"  "OTU_18414" "OTU_18422" "OTU_18428" "OTU_18445"
#> [1177] "OTU_18450" "OTU_18463" "OTU_18465" "OTU_18468" "OTU_18470" "OTU_18475"
#> [1183] "OTU_18477" "OTU_18487" "OTU_18490" "OTU_18495" "OTU_18502" "OTU_18542"
#> [1189] "OTU_18548" "OTU_18554" "OTU_18556" "OTU_18565" "OTU_18567" "OTU_18578"
#> [1195] "OTU_18579" "OTU_18580" "OTU_18588" "OTU_18593" "OTU_18594" "OTU_18598"
#> [1201] "OTU_18602" "OTU_18617" "OTU_18631" "OTU_18653" "OTU_18655" "OTU_18661"
#> [1207] "OTU_18662" "OTU_18666" "OTU_18672" "OTU_18685" "OTU_18687" "OTU_18697"
#> [1213] "OTU_18707" "OTU_18708" "OTU_18720" "OTU_1873"  "OTU_18734" "OTU_18745"
#> [1219] "OTU_18752" "OTU_18756" "OTU_18761" "OTU_18778" "OTU_18797" "OTU_18814"
#> [1225] "OTU_18815" "OTU_18819" "OTU_18830" "OTU_18840" "OTU_18844" "OTU_18855"
#> [1231] "OTU_18863" "OTU_18865" "OTU_18867" "OTU_18868" "OTU_18873" "OTU_18877"
#> [1237] "OTU_18879" "OTU_18886" "OTU_1889"  "OTU_18891" "OTU_18898" "OTU_18910"
#> [1243] "OTU_18916" "OTU_18923" "OTU_18937" "OTU_18945" "OTU_18952" "OTU_18954"
#> [1249] "OTU_18968" "OTU_18976" "OTU_18979" "OTU_1899"  "OTU_18996" "OTU_1902" 
#> [1255] "OTU_19020" "OTU_19025" "OTU_19029" "OTU_19036" "OTU_19041" "OTU_19042"
#> [1261] "OTU_19046" "OTU_19047" "OTU_19049" "OTU_19059" "OTU_19070" "OTU_19075"
#> [1267] "OTU_1910"  "OTU_19100" "OTU_19105" "OTU_19108" "OTU_19114" "OTU_19122"
#> [1273] "OTU_19124" "OTU_19160" "OTU_19170" "OTU_19189" "OTU_19214" "OTU_19215"
#> [1279] "OTU_19221" "OTU_19222" "OTU_19224" "OTU_19227" "OTU_19228" "OTU_19230"
#> [1285] "OTU_19235" "OTU_19241" "OTU_19245" "OTU_19261" "OTU_19274" "OTU_19276"
#> [1291] "OTU_19288" "OTU_19297" "OTU_19298" "OTU_19303" "OTU_19330" "OTU_19333"
#> [1297] "OTU_19335" "OTU_19347" "OTU_19348" "OTU_19362" "OTU_19371" "OTU_19378"
#> [1303] "OTU_19381" "OTU_19383" "OTU_19394" "OTU_19396" "OTU_194"   "OTU_19405"
#> [1309] "OTU_19414" "OTU_1942"  "OTU_19421" "OTU_19423" "OTU_19437" "OTU_19482"
#> [1315] "OTU_19484" "OTU_19498" "OTU_19500" "OTU_19506" "OTU_19511" "OTU_19523"
#> [1321] "OTU_19533" "OTU_19562" "OTU_19567" "OTU_19576" "OTU_19602" "OTU_1962" 
#> [1327] "OTU_19643" "OTU_19644" "OTU_19656" "OTU_1966"  "OTU_19672" "OTU_19673"
#> [1333] "OTU_19681" "OTU_19691" "OTU_19692" "OTU_19698" "OTU_19699" "OTU_1970" 
#> [1339] "OTU_19706" "OTU_1971"  "OTU_19720" "OTU_19723" "OTU_19728" "OTU_19740"
#> [1345] "OTU_19752" "OTU_19758" "OTU_19769" "OTU_19778" "OTU_19782" "OTU_19806"
#> [1351] "OTU_19807" "OTU_19814" "OTU_19825" "OTU_19827" "OTU_19832" "OTU_19858"
#> [1357] "OTU_19859" "OTU_19864" "OTU_19889" "OTU_19893" "OTU_19899" "OTU_19909"
#> [1363] "OTU_19916" "OTU_19919" "OTU_19937" "OTU_19939" "OTU_19950" "OTU_19952"
#> [1369] "OTU_19953" "OTU_19960" "OTU_19969" "OTU_19986" "OTU_19992" "OTU_20023"
#> [1375] "OTU_20024" "OTU_20025" "OTU_20026" "OTU_20027" "OTU_20038" "OTU_20046"
#> [1381] "OTU_20049" "OTU_20052" "OTU_20067" "OTU_20073" "OTU_20081" "OTU_20082"
#> [1387] "OTU_20094" "OTU_20107" "OTU_20109" "OTU_20120" "OTU_20135" "OTU_2031" 
#> [1393] "OTU_2055"  "OTU_2056"  "OTU_2062"  "OTU_2078"  "OTU_2108"  "OTU_2113" 
#> [1399] "OTU_2160"  "OTU_2168"  "OTU_2173"  "OTU_2175"  "OTU_2179"  "OTU_2181" 
#> [1405] "OTU_220"   "OTU_2212"  "OTU_2239"  "OTU_2246"  "OTU_2279"  "OTU_2283" 
#> [1411] "OTU_229"   "OTU_2309"  "OTU_2310"  "OTU_232"   "OTU_2324"  "OTU_2325" 
#> [1417] "OTU_2329"  "OTU_2343"  "OTU_2346"  "OTU_2373"  "OTU_2379"  "OTU_2398" 
#> [1423] "OTU_2406"  "OTU_2433"  "OTU_2465"  "OTU_247"   "OTU_2478"  "OTU_2481" 
#> [1429] "OTU_2483"  "OTU_2486"  "OTU_2489"  "OTU_2513"  "OTU_2527"  "OTU_2544" 
#> [1435] "OTU_255"   "OTU_2575"  "OTU_2608"  "OTU_2635"  "OTU_264"   "OTU_2647" 
#> [1441] "OTU_267"   "OTU_2676"  "OTU_2677"  "OTU_2679"  "OTU_268"   "OTU_2708" 
#> [1447] "OTU_2719"  "OTU_2723"  "OTU_273"   "OTU_2736"  "OTU_2737"  "OTU_2740" 
#> [1453] "OTU_2765"  "OTU_2769"  "OTU_2773"  "OTU_2777"  "OTU_2785"  "OTU_2791" 
#> [1459] "OTU_2793"  "OTU_2820"  "OTU_2849"  "OTU_2873"  "OTU_2887"  "OTU_2896" 
#> [1465] "OTU_2898"  "OTU_2901"  "OTU_2931"  "OTU_2946"  "OTU_2975"  "OTU_2981" 
#> [1471] "OTU_2989"  "OTU_2996"  "OTU_300"   "OTU_3003"  "OTU_3011"  "OTU_3015" 
#> [1477] "OTU_3021"  "OTU_304"   "OTU_3042"  "OTU_3045"  "OTU_3049"  "OTU_307"  
#> [1483] "OTU_3077"  "OTU_309"   "OTU_3093"  "OTU_3094"  "OTU_310"   "OTU_3106" 
#> [1489] "OTU_3117"  "OTU_3132"  "OTU_3138"  "OTU_3140"  "OTU_3149"  "OTU_3152" 
#> [1495] "OTU_3160"  "OTU_317"   "OTU_3183"  "OTU_3189"  "OTU_319"   "OTU_3194" 
#> [1501] "OTU_3197"  "OTU_321"   "OTU_3214"  "OTU_3242"  "OTU_3251"  "OTU_3253" 
#> [1507] "OTU_3256"  "OTU_3257"  "OTU_326"   "OTU_3261"  "OTU_3276"  "OTU_3278" 
#> [1513] "OTU_3286"  "OTU_329"   "OTU_3291"  "OTU_3300"  "OTU_331"   "OTU_3316" 
#> [1519] "OTU_3327"  "OTU_333"   "OTU_3341"  "OTU_3344"  "OTU_3349"  "OTU_3351" 
#> [1525] "OTU_3359"  "OTU_3388"  "OTU_3390"  "OTU_3401"  "OTU_3427"  "OTU_3438" 
#> [1531] "OTU_3448"  "OTU_3451"  "OTU_3469"  "OTU_3483"  "OTU_3497"  "OTU_3498" 
#> [1537] "OTU_3503"  "OTU_351"   "OTU_352"   "OTU_3520"  "OTU_3528"  "OTU_354"  
#> [1543] "OTU_3550"  "OTU_3613"  "OTU_3615"  "OTU_3616"  "OTU_3617"  "OTU_3630" 
#> [1549] "OTU_3642"  "OTU_3676"  "OTU_3683"  "OTU_3711"  "OTU_3728"  "OTU_3734" 
#> [1555] "OTU_3751"  "OTU_376"   "OTU_3763"  "OTU_377"   "OTU_3770"  "OTU_3779" 
#> [1561] "OTU_3800"  "OTU_3802"  "OTU_3821"  "OTU_3839"  "OTU_3854"  "OTU_3858" 
#> [1567] "OTU_3862"  "OTU_3864"  "OTU_3877"  "OTU_3891"  "OTU_3899"  "OTU_3915" 
#> [1573] "OTU_392"   "OTU_3923"  "OTU_3936"  "OTU_3940"  "OTU_3972"  "OTU_3980" 
#> [1579] "OTU_4001"  "OTU_4003"  "OTU_4035"  "OTU_4058"  "OTU_4083"  "OTU_4097" 
#> [1585] "OTU_4100"  "OTU_4104"  "OTU_413"   "OTU_4130"  "OTU_4142"  "OTU_4158" 
#> [1591] "OTU_4172"  "OTU_4183"  "OTU_4187"  "OTU_419"   "OTU_4221"  "OTU_4225" 
#> [1597] "OTU_4233"  "OTU_4248"  "OTU_427"   "OTU_4275"  "OTU_4277"  "OTU_4286" 
#> [1603] "OTU_4309"  "OTU_4312"  "OTU_4323"  "OTU_4329"  "OTU_4337"  "OTU_4348" 
#> [1609] "OTU_4356"  "OTU_4362"  "OTU_4367"  "OTU_437"   "OTU_4378"  "OTU_4381" 
#> [1615] "OTU_4382"  "OTU_4384"  "OTU_441"   "OTU_442"   "OTU_4450"  "OTU_4476" 
#> [1621] "OTU_4507"  "OTU_4516"  "OTU_4517"  "OTU_4521"  "OTU_4555"  "OTU_4584" 
#> [1627] "OTU_4589"  "OTU_4604"  "OTU_4619"  "OTU_4631"  "OTU_4653"  "OTU_4662" 
#> [1633] "OTU_4665"  "OTU_467"   "OTU_4676"  "OTU_4678"  "OTU_4695"  "OTU_4699" 
#> [1639] "OTU_4712"  "OTU_4724"  "OTU_474"   "OTU_4748"  "OTU_477"   "OTU_4788" 
#> [1645] "OTU_4794"  "OTU_48"    "OTU_4820"  "OTU_4837"  "OTU_4843"  "OTU_485"  
#> [1651] "OTU_487"   "OTU_4878"  "OTU_4896"  "OTU_4897"  "OTU_4900"  "OTU_4905" 
#> [1657] "OTU_4932"  "OTU_4935"  "OTU_4963"  "OTU_4986"  "OTU_4991"  "OTU_50"   
#> [1663] "OTU_5001"  "OTU_5011"  "OTU_5014"  "OTU_5030"  "OTU_5031"  "OTU_5042" 
#> [1669] "OTU_5043"  "OTU_506"   "OTU_5091"  "OTU_5093"  "OTU_5098"  "OTU_5121" 
#> [1675] "OTU_5130"  "OTU_5131"  "OTU_5138"  "OTU_5140"  "OTU_5150"  "OTU_5160" 
#> [1681] "OTU_5167"  "OTU_5174"  "OTU_5182"  "OTU_5193"  "OTU_522"   "OTU_5226" 
#> [1687] "OTU_5227"  "OTU_5234"  "OTU_5238"  "OTU_5240"  "OTU_5278"  "OTU_5300" 
#> [1693] "OTU_5307"  "OTU_5324"  "OTU_5327"  "OTU_5329"  "OTU_5345"  "OTU_5365" 
#> [1699] "OTU_5393"  "OTU_54"    "OTU_5418"  "OTU_5419"  "OTU_5432"  "OTU_5459" 
#> [1705] "OTU_5467"  "OTU_5471"  "OTU_5475"  "OTU_551"   "OTU_5519"  "OTU_5529" 
#> [1711] "OTU_5532"  "OTU_5541"  "OTU_5543"  "OTU_5561"  "OTU_5584"  "OTU_559"  
#> [1717] "OTU_56"    "OTU_5611"  "OTU_5627"  "OTU_5635"  "OTU_5670"  "OTU_569"  
#> [1723] "OTU_5720"  "OTU_573"   "OTU_576"   "OTU_580"   "OTU_5802"  "OTU_5812" 
#> [1729] "OTU_5826"  "OTU_583"   "OTU_5846"  "OTU_5849"  "OTU_5853"  "OTU_5909" 
#> [1735] "OTU_5926"  "OTU_5931"  "OTU_5937"  "OTU_5952"  "OTU_5964"  "OTU_5990" 
#> [1741] "OTU_600"   "OTU_6011"  "OTU_6016"  "OTU_6024"  "OTU_6051"  "OTU_6055" 
#> [1747] "OTU_6064"  "OTU_6081"  "OTU_6089"  "OTU_609"   "OTU_6095"  "OTU_6101" 
#> [1753] "OTU_612"   "OTU_6122"  "OTU_6125"  "OTU_6142"  "OTU_6144"  "OTU_6146" 
#> [1759] "OTU_618"   "OTU_6181"  "OTU_6189"  "OTU_6206"  "OTU_6219"  "OTU_624"  
#> [1765] "OTU_6251"  "OTU_6254"  "OTU_6258"  "OTU_631"   "OTU_6348"  "OTU_636"  
#> [1771] "OTU_6361"  "OTU_6375"  "OTU_6395"  "OTU_6416"  "OTU_642"   "OTU_645"  
#> [1777] "OTU_6454"  "OTU_6457"  "OTU_6487"  "OTU_6510"  "OTU_6520"  "OTU_6528" 
#> [1783] "OTU_6529"  "OTU_6563"  "OTU_657"   "OTU_6589"  "OTU_6593"  "OTU_6595" 
#> [1789] "OTU_6606"  "OTU_6608"  "OTU_6617"  "OTU_663"   "OTU_6630"  "OTU_6632" 
#> [1795] "OTU_6652"  "OTU_6671"  "OTU_6688"  "OTU_6692"  "OTU_67"    "OTU_6709" 
#> [1801] "OTU_6722"  "OTU_6745"  "OTU_6752"  "OTU_6797"  "OTU_68"    "OTU_6802" 
#> [1807] "OTU_6813"  "OTU_6815"  "OTU_6819"  "OTU_682"   "OTU_6824"  "OTU_6827" 
#> [1813] "OTU_6830"  "OTU_685"   "OTU_6865"  "OTU_6869"  "OTU_6875"  "OTU_6876" 
#> [1819] "OTU_6893"  "OTU_692"   "OTU_6920"  "OTU_6925"  "OTU_6936"  "OTU_695"  
#> [1825] "OTU_6970"  "OTU_6975"  "OTU_700"   "OTU_7010"  "OTU_702"   "OTU_7021" 
#> [1831] "OTU_704"   "OTU_7054"  "OTU_7077"  "OTU_7099"  "OTU_710"   "OTU_7114" 
#> [1837] "OTU_7129"  "OTU_713"   "OTU_7155"  "OTU_7173"  "OTU_7215"  "OTU_7232" 
#> [1843] "OTU_7247"  "OTU_7255"  "OTU_7260"  "OTU_727"   "OTU_7275"  "OTU_7287" 
#> [1849] "OTU_7293"  "OTU_7299"  "OTU_7315"  "OTU_733"   "OTU_7340"  "OTU_7342" 
#> [1855] "OTU_7365"  "OTU_7372"  "OTU_7386"  "OTU_7408"  "OTU_7416"  "OTU_7435" 
#> [1861] "OTU_7438"  "OTU_7440"  "OTU_7471"  "OTU_7480"  "OTU_7481"  "OTU_7514" 
#> [1867] "OTU_7527"  "OTU_7536"  "OTU_7554"  "OTU_7555"  "OTU_7558"  "OTU_7562" 
#> [1873] "OTU_7569"  "OTU_7581"  "OTU_760"   "OTU_7615"  "OTU_7625"  "OTU_7647" 
#> [1879] "OTU_7677"  "OTU_768"   "OTU_7689"  "OTU_7702"  "OTU_7712"  "OTU_7717" 
#> [1885] "OTU_773"   "OTU_7730"  "OTU_7731"  "OTU_7744"  "OTU_7745"  "OTU_7748" 
#> [1891] "OTU_777"   "OTU_778"   "OTU_7800"  "OTU_7820"  "OTU_7821"  "OTU_7851" 
#> [1897] "OTU_7859"  "OTU_7868"  "OTU_7885"  "OTU_7897"  "OTU_7907"  "OTU_7956" 
#> [1903] "OTU_7957"  "OTU_7959"  "OTU_7961"  "OTU_7962"  "OTU_7967"  "OTU_7990" 
#> [1909] "OTU_80"    "OTU_8026"  "OTU_8032"  "OTU_8041"  "OTU_8047"  "OTU_8055" 
#> [1915] "OTU_8057"  "OTU_8065"  "OTU_8075"  "OTU_8098"  "OTU_812"   "OTU_8134" 
#> [1921] "OTU_8137"  "OTU_8139"  "OTU_8180"  "OTU_821"   "OTU_8242"  "OTU_8244" 
#> [1927] "OTU_8268"  "OTU_8277"  "OTU_8305"  "OTU_8307"  "OTU_8309"  "OTU_8316" 
#> [1933] "OTU_8333"  "OTU_8342"  "OTU_8344"  "OTU_8373"  "OTU_8404"  "OTU_8408" 
#> [1939] "OTU_8455"  "OTU_8518"  "OTU_8523"  "OTU_8562"  "OTU_8574"  "OTU_8576" 
#> [1945] "OTU_8594"  "OTU_8600"  "OTU_8660"  "OTU_8769"  "OTU_878"   "OTU_880"  
#> [1951] "OTU_8809"  "OTU_881"   "OTU_8814"  "OTU_8839"  "OTU_8850"  "OTU_8855" 
#> [1957] "OTU_886"   "OTU_8888"  "OTU_8897"  "OTU_8902"  "OTU_8911"  "OTU_8915" 
#> [1963] "OTU_8939"  "OTU_8947"  "OTU_8954"  "OTU_8977"  "OTU_8980"  "OTU_8988" 
#> [1969] "OTU_900"   "OTU_9003"  "OTU_9005"  "OTU_9039"  "OTU_9086"  "OTU_9099" 
#> [1975] "OTU_9106"  "OTU_9112"  "OTU_9113"  "OTU_9150"  "OTU_9186"  "OTU_9191" 
#> [1981] "OTU_9229"  "OTU_9256"  "OTU_9294"  "OTU_9309"  "OTU_9312"  "OTU_9326" 
#> [1987] "OTU_9332"  "OTU_9339"  "OTU_9380"  "OTU_9412"  "OTU_9425"  "OTU_9460" 
#> [1993] "OTU_9466"  "OTU_9506"  "OTU_9508"  "OTU_9512"  "OTU_9518"  "OTU_9556" 
#> [1999] "OTU_9570"  "OTU_9590"  "OTU_9599"  "OTU_9607"  "OTU_9610"  "OTU_9614" 
#> [2005] "OTU_963"   "OTU_9647"  "OTU_9648"  "OTU_9688"  "OTU_9690"  "OTU_9712" 
#> [2011] "OTU_9720"  "OTU_9733"  "OTU_9742"  "OTU_9754"  "OTU_9774"  "OTU_9793" 
#> [2017] "OTU_9796"  "OTU_9805"  "OTU_9809"  "OTU_9818"  "OTU_9833"  "OTU_9838" 
#> [2023] "OTU_9848"  "OTU_985"   "OTU_9862"  "OTU_9876"  "OTU_9892"  "OTU_9894" 
#> [2029] "OTU_9906"  "OTU_9908"  "OTU_9926"  "OTU_9929"  "OTU_9936"  "OTU_9977" 
#> [2035] "OTU_9984"  "OTU_9996"