Calculate the mean, variance and CV of each OTU

Calculate the mean, variance and CV of each OTU

summarise_taxa(otu_table, taxa_as_rows = TRUE)

Arguments

otu_table

a dataframe of OTUs where the first row is the OTU ID and column names refer to sites

taxa_as_rows

logical data must be in a format where the taxa are rows and the sites are columns. The default value is TRUE, if FALSE data will be transposed for downstream analysis.

Value

a dataframe with the Mean, Variance and CV for each OTU

Examples

summarise_taxa(arabidopsis)
#> # A tibble: 14,890 × 4
#>    X            Mean Variance    CV
#>    <chr>       <dbl>    <dbl> <dbl>
#>  1 OTU_100   0.368    0.707   1.92 
#>  2 OTU_1000  0.00694  0.00692 0.997
#>  3 OTU_10000 0.698    5.85    8.38 
#>  4 OTU_10002 0.260    0.284   1.09 
#>  5 OTU_10003 0.00694  0.00692 0.997
#>  6 OTU_10006 0.146    0.160   1.10 
#>  7 OTU_10007 0.00694  0.00692 0.997
#>  8 OTU_10008 0.0451   0.0502  1.11 
#>  9 OTU_1001  0.0694   0.128   1.84 
#> 10 OTU_10011 0.861    5.30    6.16 
#> # … with 14,880 more rows